Work in Progress (updating UI with Peer Feedback)

+++ src/main/java/nuggets/IdleItem.java
@@ -2,10 +2,10 @@ package nuggets;
 
 public class IdleItem{
     public int itemInd; // index of item
-    public int itemPrice; // price of item
+    public double itemPrice; // price of item
     public int itemProgress; // how many times has this item been 
-    public int itemProgRate; // how fast the item scales (Linear)
-    public int itemPriceRate; // rate of price scale (exoponential)
+    public double itemProgRate; // how fast the item scales (Linear)
+    public double itemPriceRate; // rate of price scale (exoponential)
     public int itemInit; // starting value of the item, example: item gives itemInit amount of money generation when bought
     public String itemName; // name of item
 
@@ -13,7 +13,7 @@ public class IdleItem{
     public static int UCearned = 1;
     public static int UCpassive = 300000;
 
-    public IdleItem(int index, int price, String name, int prog, int progRate, int priceRate, int init){
+    public IdleItem(int index, double price, String name, int prog, double progRate, double priceRate, int init){
         itemInd = index;
         itemPrice = price;
         itemName = name;
